requirements.txt
python-dotenv
bottle
tenacity
